backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pydantic-settings>=2.2,<3.0
psycopg[binary]>=3.1,<4.0
celery[redis]>=5.4,<6.0
alembic>=1.13,<2.0
redis>=5.0,<6.0
httpx>=0.27,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib>=1.7,<2.0
python-multipart>=0.0.9,<1.0
